backend/requirements.txt
flask
flask-cors
flask-jwt-extended
bcrypt
python-dotenv
groq